Move sort functions into their own file. Export necessary sort functions into utils.ts for sortItems() utility function
